feat: integ tests for jumpstart model and estimator
