+++ tests/integ/sagemaker/jumpstart/constants.py
